+++ app/controllers/accounts_controller.rb
@@ -1,20 +1,14 @@
 class AccountsController < ApplicationController
   before_filter :authenticate_user!
+  before_filter :set_account, only: [:show, :edit, :update, :destroy]
 
   def index
-    # current_user.update_attribute :accounts_per_page, params[:accounts_per_page].to_i if params[:accounts_per_page].to_i >= 10 and params[:accounts_per_page].to_i <= 200
-    account_scope = current_user.accounts
-    @search_terms = params[:search].to_s.gsub(/[^0-9a-zA-Z]/, ' ').split(' ')
-    @search_terms.each{|search_term| account_scope = account_scope.search(search_term) }
-
     @order = scrub_order(Account, params[:order], 'accounts.name')
-    account_scope = account_scope.order(@order)
-
-    @accounts = account_scope.page(params[:page]).per(20) # current_user.accounts_per_page)
+    @accounts = current_user.accounts.search(params[:search]).order(@order).page(params[:page]).per(20)
   end
 
   def show
-    @account = current_user.accounts.find(params[:id])
+
   end
 
   def new
@@ -22,32 +16,28 @@ class AccountsController < ApplicationController
   end
 
   def edit
-    @account = current_user.accounts.find(params[:id])
+
   end
 
   def create
     @account = current_user.accounts.new(post_params)
     if @account.save
-      flash[:notice] = 'Account was successfully created.'
-      redirect_to @account
+      redirect_to @account, notice: 'Account was successfully created.'
     else
       render action: :new
     end
   end
 
   def update
-    @account = current_user.accounts.find(params[:id])
     if @account.update_attributes(post_params)
-      flash[:notice] = 'Account was successfully updated.'
-      redirect_to @account
+      redirect_to @account, notice: 'Account was successfully updated.'
     else
       render action: :edit
     end
   end
 
   def destroy
-    @account = current_user.accounts.find(params[:id])
-    @account.destroy
+    @account.destroy if @account
     redirect_to accounts_path
   end
 
@@ -59,6 +49,10 @@ class AccountsController < ApplicationController
     params[:account].slice(
       :name
     )
+  end
+
+  def set_account
+    @account = current_user.accounts.find_by_id(params[:id])
   end
 
 end